Move widget code to viewer module
