+++ higlass/viewer.py
@@ -1,4 +1,37 @@
-from .widgets import HiGlassDisplay
+import ipywidgets as widgets
+from traitlets import Unicode
+# from traitlets import default
+from traitlets import List
+from traitlets import Dict
+from traitlets import Int
+
+from ._version import __version__
+
+
+@widgets.register
+class HiGlassDisplay(widgets.DOMWidget):
+    _view_name = Unicode("HiGlassDisplayView").tag(sync=True)
+    _model_name = Unicode("HiGlassDisplayModel").tag(sync=True)
+    _view_module = Unicode("higlass-jupyter").tag(sync=True)
+    _model_module = Unicode("higlass-jupyter").tag(sync=True)
+    _view_module_version = Unicode(__version__).tag(sync=True)
+    _model_module_version = Unicode(__version__).tag(sync=True)
+
+    _model_data = List([]).tag(sync=True)
+    viewconf = Dict({}).tag(sync=True)
+    hg_options = Dict({}).tag(sync=True)
+    height = Int().tag(sync=True)
+
+    def __init__(self, **kwargs):
+        # self.viewconf = viewconf
+        super(HiGlassDisplay, self).__init__(**kwargs)
+
+    # @default('layout')
+    # def _default_layout(self):
+    #     return widgets.Layout(height='600px', align_self='stretch')
+
+    # def set_data(self, js_data):
+    #     self._model_data = js_data
 
 
 def display(views, location_syncs=[], zoom_syncs=[], host='localhost', server_port=None):
@@ -26,11 +59,11 @@ def display(views, location_syncs=[], zoom_syncs=[], host='localhost', server_po
         for track in view.tracks:
             if track.tracks:
                 for track1 in track.tracks:
-                    if ('server' not in track1.viewconf or 
+                    if ('server' not in track1.viewconf or
                             track1.viewconf['server'] is None):
                         track1.viewconf['server'] = server.api_address
             else:
-                if ('server' not in track.viewconf or 
+                if ('server' not in track.viewconf or
                         track.viewconf['server'] is None):
                     track.viewconf['server'] = server.api_address
